backend/requirements.txt
# Django Core
Django==5.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
django-environ==0.11.0

# Database (Development: SQLite, Production: PostgreSQL)
# psycopg2-binary==2.9.9  # PostgreSQL adapter - Comment out for development (SQLite), uncomment for production
# dj-database-url==2.1.0  # Optional: For parsing DATABASE_URL in production (install if needed)

# Cache & Task Queue (Optional for development)
redis==5.0.0
celery==5.3.0
django-celery-beat==2.5.0

# API Documentation
drf-spectacular==0.26.0

# Payment Processing
stripe==7.0.0
paypalrestsdk==1.13.0

# Notifications
twilio==8.10.0
sendgrid==6.11.0

# Calendar Integration
google-api-python-client==2.100.0
msal==1.25.0

# Utilities
Pillow==10.3.0  # Updated version to avoid build issues
python-dateutil==2.8.2
requests==2.31.0  # For Google Places API and external API calls